requirements.txt
Pillow
PyGetWindow
numpy
opencv-python
mouse
keyboard
PyAutoGUI
